fastapi/cloud/requirements.txt
# FastAPI 相关
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Nacos SDK Python 2.0.7
nacos-sdk-python==2.0.7

# 环境变量管理
python-dotenv>=1.0.0

# 其他可能需要的依赖
requests>=2.31.0
